Backend/requirements.txt
django~=3.2
djangorestframework~=3.14.0
django-filter~=23.2